move out player component to improve performance
